♻️ refactor: api 함수 정리
